Wrote logs for every movement done.

+++ app.py
@@ -1,7 +1,5 @@
 from flask import Flask, render_template, request, redirect, url_for, session
 from flask_mysqldb import MySQL
-#from flaskext.mysql import MySQL
-#import pymysql
 import yaml
 import bcrypt
 
@@ -19,13 +17,12 @@ mysql = MySQL(app)
 # ANGAJATI DONE
 @app.route('/angajati', methods=['GET', 'POST'])
 def getAngajati():
-    print("___app.py/home::angajati: STARTED")
-    if request.method=='GET':
+    print("___app.py / getAngajati::angajati.html: Start getting data from DB.")
+    if request.method == 'GET':
         if session.get('username') is None:
-            print("Nu e logat")
+            print("___app.py / getAngajati::angajati.html: Not logged in. Redirect to login.")
             return redirect(url_for('login'))
         if session['username']:
-            #return render_template('angajati.html')
             try:
                 curs = mysql.connection.cursor()
                 curs.execute('SELECT * FROM tblAngajati')
@@ -33,15 +30,15 @@ def getAngajati():
                 # pentru a vedea atributele unui obiect:
                 #for ang in data:
                  #   print(ang['idAngajat'])
-                print('___app.py / home::angajati: FETCH DONE.')
+                print('___app.py / getAngajati::angajati.html: FETCH DONE.')
                 curs.close()
             except Exception as e:
-                print('ERROR___app.py / home::angajati: EROARE FETCH____', e)
+                print('ERROR___app.py / getAngajati::angajati.html: EROARE FETCH____', e)
 
             return render_template('angajati.html', angajati=data)
 
 
-
+# NOT DONE YET
 @app.route('/adauga', methods=['GET', 'POST'])
 def adauga():
     if request.method == 'POST':
@@ -97,7 +94,7 @@ def editAngajat(id):
     curs = mysql.connection.cursor()
     if request.method == "GET":
         if session.get('username') is None:
-            print("Nu e logat")
+            print("___app.py / editAngajat::editAngajat.html: Not logged in. Redirect to login.")
             return redirect(url_for('login'))
         if session['username']:
             sql = 'SELECT * FROM tblAngajati WHERE idAngajat=%s'
@@ -121,14 +118,11 @@ def editAngajat(id):
             curs.execute(sql, [numeToUpdate, prenumeToUpdate, functieToUpdate, dataAngToUpd, telefonToUpdate, emailToUpdate, salariuToUpdate, idToUpdate])
             mysql.connection.commit()
             curs.close()
+            print("___app.py / editAngajat::editAngajat.html: Update done.")
             return redirect(url_for('getAngajati'))
         except Exception as e:
+            print("ERROR___app.py / editAngajat::editAngajat.html: ")
             print(e)
-    # sql = 'SELECT * FROM tblAngajati WHERE idAngajat=%s'
-    # curs.execute(sql,[id])
-    # res = curs.fetchone()
-    # curs.close()
-    # return render_template("editAngajat.html", angajati=res)
 
 # INCERCARE
 def delete(numeidElem,tabel,valoareId):
@@ -139,24 +133,37 @@ def delete(numeidElem,tabel,valoareId):
     curs.close()
 # END INCERCARE
 
+
+# MAI E PUTIN;
 @app.route('/deleteAngajat/<string:id>', methods=['GET','POST'])
 def deleteAngajat(id):
-        curs = mysql.connection.cursor()
-        sql = "DELETE FROM tblAngajati WHERE idAngajat=%s"
-        curs.execute(sql, [id])
-        mysql.connection.commit()
-        curs.close()
-
-        return redirect(url_for('getAngajati'))
+    curs = mysql.connection.cursor()
+    if request.method == "GET":
+        if session.get('username') is None:
+            print("___app.py / deleteAngajat::deleteAngajat.html: Not logged in. Redirect to login.")
+            return redirect(url_for('login'))
+        if session['username']:
+            #curs = mysql.connection.cursor()
+            sql = "DELETE FROM tblAngajati WHERE idAngajat=%s"
+            curs.execute(sql, [id])
+            mysql.connection.commit()
+            curs.close()
+            print("___app.py / deleteAngajat::deleteAngajat.html: idAngajat %s deleted." % id)
+            return redirect(url_for('getAngajati'))
 
 
 @app.route('/register', methods=['GET', 'POST'])
 def register():
-    if request.method == 'GET':
-        return render_template('register.html')
-    else:
-        session['reg'] = ""
+    if request.method =='GET':
+        if session.get('username') is None:
+            print("___app.py / register::register.html: Session clear. Not logged in. Proceed to register.")
+            return render_template('register.html')
+        if session['username']:
+            print("___app.py / register::register.html: User %s logged in. Redirect to homePage." % session.__getitem__('username'))
+            return redirect(url_for('homePage'))
 
+    if request.method == 'POST':
+        session['reg'] = ""
         numeN        = request.form['nume']
         prenumeN     = request.form['prenume']
         emailN       = request.form['email']
@@ -172,13 +179,18 @@ def register():
             eroareRegister = "OK"
             session['reg']='OK'
             #return 'register success'
+            print("___app.py / register::register.html: Register success. Username: %s." % [userN])
             session.clear()
+            print("___app.py / register::register.html: CLEARING SESSION. PROCEED TO LOGIN. ")
             return redirect(url_for('login'))
             #return render_template('register.html', eroareHTML=eroareRegister)
         except Exception as e:
+            print("___app.py / register::register.html: Error down.")
             print(e)
             if e.args[0] == 1062:
                 eroareRegister = "Username already exist. Try another."
+                print("___app.py / register::register.html: Username already exists in table.")
+
                 session['reg'] = '1062'
                 #return render_template('register.html', eroareHTML='MUIE')
                 #return 'username already exists'
@@ -191,11 +203,12 @@ def register():
 
 @app.route('/homePage', methods=['GET', 'POST'])
 def homePage():
-    if request.method=="GET":
+    if request.method == "GET":
         if session.get('username') is None:
-            print("Nu e logat")
+            print("___app.py / homePage::homePage.html: User not logged in. Redirect to login.")
             return redirect(url_for('login'))
         if session['username']:
+            print('___app.py / homePage::homePage.html: USER %s LOGGED IN.' % session.__getitem__('username') )
             return render_template('homePage.html')
     # if request.method == 'GET':
     #     return render_template('homePage.html')
@@ -226,15 +239,15 @@ def homePage():
     #             #return 'login failed'
 
 
-
 @app.route('/login', methods=['GET', 'POST'])
 def login():
     if request.method == 'GET':
 
         if session.get('username') is None:
-            print("Nu e logat")
+            print("___app.py / login::login.html: User not logged in.")
             return render_template('login.html')
         if session['username']:
+            print("___app.py / login::login.html: User %s logged in. Redirect to homePage." % session.__getitem__('username'))
             return redirect(url_for('homePage'))
         #else:
             #return render_template('login.html')
@@ -254,13 +267,16 @@ def login():
         if len(userLogin) > 0:
             if userInfo is None:
                 session['authenticated'] = False
+                print("___app.py / login::login.html: Username not found.")
                 return redirect(url_for('login'))
             if bcrypt.hashpw(passLogin, userInfo['Password'].encode('utf-8')) == userInfo['Password'].encode('utf-8'):
                 session['username'] = userInfo['Username']
                 session['authenticated'] = True
+                print("___app.py / login::login.html: Login success. Username: %s. Redirect to homePage." % session.__getitem__('username'))
                 return redirect(url_for('homePage'))
             else:
                 session['authenticated'] = False
+                print("___app.py / login::login.html: LOGIN FAILED.")
                 return redirect(url_for('login'))
 
 
@@ -268,7 +284,9 @@ def login():
 def logout():
     #print(session.__getitem__('username'))
     #session['username'] = ""
+    print("___app.py / logout::logout.html: Log out username: %s." % session.__getitem__('username'))
     session.clear()
+    print("___app.py / logout::logout.html: Logout done. Session clear. Redirect to login.")
     #session['username']=""
     #session['authenticated'] = False
     return redirect(url_for('login'))
